Function to calculate and evaluate the NUE

This is the NUE calculation as being used in the BBWP framework

Usage

bln_nut_nue(
  B_LU_BRP,
  B_HELP_WENR,
  B_GWL_CLASS,
  A_P_AL,
  A_P_CC,
  A_P_WA,
  A_N_RT,
  B_N_RT = NA_real_,
  B_N_RT_SD = NA_real_,
  penalty = TRUE
)

Arguments

B_LU_BRP

(numeric) The crop code

B_HELP_WENR

(character) The soil type abbreviation, derived from 1:50.000 soil map

B_GWL_CLASS

(character) The groundwater table class

A_P_AL

(numeric) The P-AL content of the soil

A_P_CC

(numeric) The plant available P content, measured via 0.01M CaCl2 extraction (mg / kg)

A_P_WA

(numeric) The P-content of the soil extracted with water (mg P2O5 / 100 ml soil)

A_N_RT

(numeric) The organic nitrogen content of the soil (mg N / kg)

B_N_RT

(numeric) The mean organic nitrogen content of the soil in the LSW region (mg N / kg). Optional.

B_N_RT_SD

(numeric) The variance in organic nitrogen content of the soil in the LSW region (standard deviation) (mg N / kg). Optional.

penalty

(boolean) the option to apply a penalty for high risk BBWP field indicators. Default is TRUE.